refactor: export latest manifest, and version schemas separately
