+++ src/streamdeck/plugins/manifest/v6.4.ts
