+++ src/streamdeck/plugins/manifest/v6.6.ts
@@ -1,4 +1,4 @@
-import type { Manifest } from "../latest";
+import type { Manifest } from "./latest";
 
 /**
  * Defines the plugin and available actions, and all information associated with them, including the plugin's entry point, all iconography, action default behavior, etc.
